object-detection-api/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ML
ultralytics==8.3.40
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow==11.0.0
numpy>=1.24.0

# HTTP client (for /detect/url)
httpx==0.28.0

# Rate limiting
slowapi==0.1.9

# WebSocket client (for demo script)
websockets>=12.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
